added redux flow for <DepthChart />
