+++ package.json
@@ -9,7 +9,7 @@
     "test": "react-scripts test --env=jsdom --watch",
     "eject": "react-scripts eject",
     "flow": "flow",
-    "scss": "scss --watch src/styles/scss/storybook.scss:src/styles/css/index.css && scss --watch src/styles/scss/index.scss:src/styles/css/app.css",
+    "scss": "scss --watch src/styles/scss/storybook.scss:src/styles/css/storybook.css && scss --watch src/styles/scss/index.scss:src/styles/css/app.css",
     "storybook": "start-storybook -p 9009 -s public",
     "build-storybook": "build-storybook -s public"
   },
