+++ src/app/App/App.js
@@ -1,16 +1,8 @@
 import React, { Component } from 'react';
 import { defineMessages, FormattedMessage } from 'react-intl';
-import OHLCV from '../../components/OHLCV';
-import OrderBook from '../../components/OrderBook';
-import HomePage from '../../components/HomePage';
 
-class TestApisPage extends Component {
-  state = {
-    ohlcvData: [{ data: '' }],
-  };
-
+class App extends Component {
   render() {
-    const { ohlcvData } = this.state;
     return (
       <div className="App">
         <header className="App-header">
@@ -27,21 +19,7 @@ class TestApisPage extends Component {
   }
 }
 
-export default TestApisPage;
-
-const AppDiv = () => (
-  <div className="App">
-    <header className="App-header">
-      <img className="App-logo" alt="logo" />
-      <h1 className="App-title">
-        <FormattedMessage {...messages.title} />
-      </h1>
-    </header>
-    <p className="App-intro">
-      <FormattedMessage {...messages.introduction} values={{ code: <code>src/App.js</code> }} />
-    </p>
-  </div>
-);
+export default App;
 
 const messages = defineMessages({
   title: {
