+++ src/app/TestReduxandApis/index.js
@@ -3,6 +3,7 @@ import OHLCV from '../../components/OHLCV';
 import OrderBook from '../../components/OrderBook';
 import OrderHistory from '../../components/OrderHistory';
 import TradeHistory from '../../components/TradeHistory';
+import DepthChart from '../../components/DepthChart';
 import HomePage from '../HomePage';
 
 class TestApisPage extends Component {
@@ -10,6 +11,7 @@ class TestApisPage extends Component {
     return (
       <React.Fragment>
         <HomePage />
+        <DepthChart />
         <OrderHistory />
         <TradeHistory />
         <OHLCV />
