+++ src/components/DepthChart/stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { text, withKnobs } from '@storybook/addon-knobs/react';
 import { withInfo } from '@storybook/addon-info';
-import DepthChartRenderer from './index';
+import DepthChart from './DepthChart';
 import * as README from './README.md';
 import * as bidAsk from '../../jsons/bidAsk.json';
 
@@ -11,12 +11,12 @@ storiesOf('DepthChart', module)
   .add(
     'Loading',
     withInfo({ text: README, source: false })(() => (
-      <DepthChartRenderer data={bidAsk.list} loading={true} title="Price (BTC/USDT)" />
+      <DepthChart data={bidAsk.list} loading={true} title="Price (BTC/USDT)" />
     ))
   )
   .add(
     'Not Loading',
     withInfo({ text: README, source: false })(() => (
-      <DepthChartRenderer data={bidAsk.list} loading={false} title="Price (BTC/USDT)" />
+      <DepthChart data={bidAsk.list} loading={false} title="Price (BTC/USDT)" />
     ))
   );
