+++ src/store/actions/depthChart.js
@@ -1,0 +1,12 @@
+const actionTypes = {
+  saveData: 'depthChart/SAVE_DATA',
+};
+
+export function saveData(data: any) {
+  return {
+    type: actionTypes.saveData,
+    payload: { data },
+  };
+}
+
+export default actionTypes;
