+++ src/store/models/homePage.js
@@ -8,10 +8,12 @@ import * as ohlcvActionCreators from '../actions/ohlcv';
 import * as orderBookActionCreators from '../actions/orderBook';
 import * as tradeHistoryActionCreators from '../actions/tradeHistory';
 import * as orderHistoryActionCreators from '../actions/orderHistory';
+import * as depthChartActionCreators from '../actions/depthChart';
 
 import * as orderList from '../../jsons/ordersList.json';
 import * as tradeHistory from '../../jsons/tradeHistory.json';
 import * as orderHistory from '../../jsons/orderHistory.json';
+import * as bidAsk from '../../jsons/bidAsk.json';
 
 export default function getHomePageModel(state: State) {
   return HomePageModel(state.homePage);
@@ -38,6 +40,11 @@ const orderHistoryData = {
   decimals: 7,
   authenticated: true,
 };
+const depthChartData = {
+  data: bidAsk.list,
+  loading: false,
+  title: 'ETJ/BTC',
+};
 
 export const loadData = ({ tokenId }: LoadDataParams): ThunkAction => {
   return async (dispatch, getState) => {
@@ -47,5 +54,6 @@ export const loadData = ({ tokenId }: LoadDataParams): ThunkAction => {
     dispatch(orderBookActionCreators.saveData(orderBookData));
     dispatch(tradeHistoryActionCreators.saveData(tradeHistoryData));
     dispatch(orderHistoryActionCreators.saveData(orderHistoryData));
+    dispatch(depthChartActionCreators.saveData(depthChartData));
   };
 };
