+++ src/types/depthChart.js
@@ -1,0 +1,13 @@
+//@flow
+export type Order = {
+  time: number,
+  type: string,
+  amount: number,
+  price: number,
+};
+
+export type DepthChartState = {
+  data: Array<Object>,
+  loading: boolean,
+  title: string,
+};
